Add post-processing for int overflow bug

+++ Octokit.GraphQL/Model/CheckRun.cs
@@ -42,7 +42,7 @@ namespace Octokit.GraphQL.Model
         /// <summary>
         /// Identifies the primary key from the database.
         /// </summary>
-        public int? DatabaseId { get; }
+        public long? DatabaseId { get; }
 
         /// <summary>
         /// The corresponding deployment for this job, if any
